refactor: set util
